+++ DietCalculatorSystem/Services/Diets/DietService.cs
@@ -194,7 +194,7 @@ namespace DietCalculatorSystem.Services.Diets
         public void CalculateDiet(string name,
             int? age,
             string gender,
-            int? height,
+            double? height,
             double? weight,
             double? activity)
         {
